server/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "sequence"
version = "0.1.0"
description = "Sequence package."
requires-python = ">=3.10"
dependencies = [
    "python-dotenv==1.2.1",
    "openai==2.24.0",
    "redis==7.2.1",
    "asyncpg==0.31.0",
    "fastapi==0.135.1",
    "uvicorn==0.41.0",
    "python-multipart==0.0.22",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["sequence", "sequence.*"]
